backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==4.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
frozenlist==1.8.0
gunicorn==21.2.0
h11==0.16.0
idna==3.11
jwt==1.4.0
Mako==1.3.10
MarkupSafe==3.0.3
migrate==0.3.8
multidict==6.7.0
passlib==1.7.4
propcache==0.4.1
psycopg2-binary>=2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
twilio==9.8.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.22.0
